Add ReportsService from ParsingService
