+++ HW8/Valeriy_Satkov/MultithreadApp/Multithread.Core/Multithread.Core.csproj
@@ -50,6 +50,7 @@
     <Compile Include="Services\FileManager.cs" />
     <Compile Include="Interfaces\IFileManager.cs" />
     <Compile Include="Services\LoadService.cs" />
+    <Compile Include="Services\ReportsService.cs" />
     <Compile Include="Services\ParsingService.cs" />
   </ItemGroup>
   <ItemGroup>
